+++ src/utils/group/index.d.ts
@@ -1,7 +1,3 @@
-export interface NormalizedTreeValue {
-  parent: any;
-}
-
 export interface TreeValue {
   children?: TreeValue[];
 }
@@ -20,18 +16,8 @@ export type Group = string | {
   }
 }
 
-type GroupValues = { [path: string]: Array<any> }
+export type GroupValues = { [path: string]: Array<any> }
 
-export interface BuildTreeOptions {
-  /** Default is 'id' */
-  idPath?: string;
-  /** Default is 'parent' */
-  parentPath?: string;
-  /** How id and parent should be compared. If not provided then strict '===' will be used */
-  comparator?: CompareCallback
-}
-/** Builds a tree of normalized array values */
-export declare function buildTree(array: NormalizedTreeValue[], options?: BuildTreeOptions): Promise<TreeValue[]>
 /** Extracts unique values for each group */
 export declare function extractGroupValues(array: Object[], groups: Group[]): Promise<GroupValues>
 export declare function buildGroupsTree(groupValues: GroupValues): Promise<TreeValue[]>
